use size_t for pcm_sizeOfConvBuffer() result
In my previous patch set, I forgot to change the
pcm_sizeOfConvBuffer() invocation in convertAudioFormat() to also use
size_t.

git-svn-id: https://svn.musicpd.org/mpd/trunk@7292 09075e82-0dd4-0310-85a5-a0d7c8717e4f

+++ src/audioOutput.c
@@ -184,11 +184,11 @@ int openAudioOutput(AudioOutput * audioOutput, AudioFormat * audioFormat)
 static void convertAudioFormat(AudioOutput * audioOutput, char **chunkArgPtr,
 			       int *sizeArgPtr)
 {
-	int size = pcm_sizeOfConvBuffer(&(audioOutput->inAudioFormat),
-                                        *sizeArgPtr,
-                                        &(audioOutput->outAudioFormat));
+	size_t size = pcm_sizeOfConvBuffer(&(audioOutput->inAudioFormat),
+					   *sizeArgPtr,
+					   &(audioOutput->outAudioFormat));
 
-	if (size > audioOutput->convBufferLen) {
+	if (size > (size_t)audioOutput->convBufferLen) {
 		audioOutput->convBuffer =
 		    xrealloc(audioOutput->convBuffer, size);
 		audioOutput->convBufferLen = size;
